refactor: wantplace 패키지 추출
